% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{acc}
\alias{acc}
\title{Unique combinations of variables}
\usage{
acc(k)
}
\arguments{
\item{k}{Number of covariates (or covariate groups).}
}
\description{
Function to get all possible combinations for a set of covariates.
}
\details{
This function takes a number (\code{k}) and then outputs a matrix with every
row depicting one unique combination of covariates. This function is used for
model fitting.
}
