% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{args_supported}
\alias{args_supported}
\title{Is this combination of class/gof/varimp supported by \link{vibe}?}
\usage{
args_supported(object = NULL, varimp = NULL, gof = NULL, progress = NULL)
}
\description{
This function exists to check that all arguments are properly
specified around most functions in the package, and if they are, whether
the object/gof/varimp is supported.
}
\keyword{internal}
