% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fold_change.R
\name{calc_fc}
\alias{calc_fc}
\title{Calculate the fold change and p-value for each symbol between two splitting_var groups per grouping_var group}
\usage{
calc_fc(
  df,
  threshold_method = c("mean", "median"),
  grouping_var = "intr",
  return_groups = NA,
  splitting_var,
  reverse = FALSE
)
}
\arguments{
\item{df}{dataframe harmonized by VIBE's harmonize_df function}

\item{threshold_method}{the summary stat method to use, either "mean" or "median"}

\item{grouping_var}{column name which you want to group data by (intr if left to NA)}

\item{return_groups}{Character vector with grouping_var values you wish to include in the results table}

\item{splitting_var}{column name which splits the dataset into two groups between which the fold change is calculated}

\item{reverse}{TRUE or FALSE, if FALSE fold chanae is calculated as b/a, if TRUE as a/b}
}
\value{
table with fold change and Kruskal Wallis p-value per unique primary_site and gene combination
}
\description{
Calculate the fold change and p-value for each symbol between two splitting_var groups per grouping_var group
}
\examples{
# prepare data for use with VIBE functions
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))

# example1: calculate foldchange between two databases, database2/database1
unique(df_harmonized$database)
calc_fc(df_harmonized, threshold_method = "median", splitting_var = "database")

# example2: calculate foldchange between two databases, database1/database2
unique(df_harmonized$database)
calc_fc(df_harmonized, threshold_method = "median", splitting_var = "database",
       reverse = TRUE)

# example3: group by indication instead of intr
unique(df_harmonized$database)
calc_fc(df_harmonized, threshold_method = "median", splitting_var = "database",
       grouping_var = "indication")
}
