% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fold_change_tcga_gtex.R
\name{calc_fc_gtex_tcga}
\alias{calc_fc_gtex_tcga}
\title{Calculate the fold change and p-value between GTEX and TCGA samples of the same primary site}
\usage{
calc_fc_gtex_tcga(df, threshold_method, return_primary_site = NA)
}
\arguments{
\item{df}{dataframe harmonized by VIBE's harmonize_df function, with the indication column being a concatenation of the study and primary_site columns of the XENA dataset.}

\item{threshold_method}{the summary stat method to use, either "mean" or "median"}

\item{return_primary_site}{Character vector with primary_site values (name of tissue column in Xena dataset) you wish to include in the results table}
}
\value{
table with fold change and Kruskal Wallis p-value per unique primary_site and gene combination
}
\description{
Calculate the fold change and p-value between GTEX and TCGA samples of the same primary site
}
