% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_percentage_per_quadrant.R
\name{calculate_percentage_per_quadrant}
\alias{calculate_percentage_per_quadrant}
\title{Calculate percentage of sample in each quadrant}
\usage{
calculate_percentage_per_quadrant(quad_summary, grouping_var = NA)
}
\arguments{
\item{quad_summary}{the wide format of the data including the quadrant information column as calculated by @add_quadrant_info function.}

\item{grouping_var}{groups object based on this variable; intr if not provided}
}
\value{
tibble with combined data_wide and quadrant information (as an additional column)
}
\description{
Calculate percentage of sample in each quadrant
}
\examples{
# prepare data
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))


# example 1
df_quadrant <- add_quadrant_info(df = df_harmonized,
                                 target1 = "Immune target",
                                 target2 = "Tumor target",
                                 threshold_method = "median")

calculate_percentage_per_quadrant(df_quadrant)

# example 2
df_quadrant <- add_quadrant_info(df = df_harmonized,
                                 target1 = "Immune target",
                                 target2 = "Tumor target",
                                 grouping_var = "indication",
                                 threshold_method = "median")

calculate_percentage_per_quadrant(df_quadrant, grouping_var = "indication")
}
