% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{class_finder}
\alias{class_finder}
\title{Class finder}
\usage{
class_finder(object)
}
\description{
This function is just there to find the class of object
and only give back one class...
}
\keyword{internal}
