% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/89_general_helpers.R
\name{det_npar}
\alias{det_npar}
\title{Number of modeled parameters determiner}
\usage{
det_npar(object)
}
\description{
Determines the number of modeled parameters in gamlss models
}
\keyword{internal}
