% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction_and_axis_quadrants.R
\name{direction_and_axis_quadrants}
\alias{direction_and_axis_quadrants}
\title{Obtain the quadrant specific direction of selected genes and returns positioning of labels on the quadrant}
\usage{
direction_and_axis_quadrants(
  df,
  QOI = c("Q1", "Q2", "Q3", "Q4"),
  target1,
  target2
)
}
\arguments{
\item{df}{Dataframe as harmonized by @harmonize_df}

\item{QOI}{Quadrant of interest among 4 quadrants}

\item{target1}{Official gene symbol of gene 1}

\item{target2}{Official gene symbol of gene 2}
}
\value{
a list comprising of the information of the direction of both the target (high/low) and the plotting location for each quadrant
}
\description{
Obtain the quadrant specific direction of selected genes and returns positioning of labels on the quadrant
}
\examples{
# prepare data
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))


# example 1: positioning of labels Q1 (samples low in target1, high in target2)
direction_and_axis_quadrants(df = df_harmonized, QOI = "Q1",
                            target1 = "Immune target", target2 = "Tumor target")
# example 2: positioning of labels Q2 (samples high in both targets)
direction_and_axis_quadrants(df = df_harmonized, QOI = "Q2",
                            target1 = "Immune target", target2 = "Tumor target")
# example 3: positioning of labels Q3 (samples high in target1, low in target 2)
direction_and_axis_quadrants(df = df_harmonized, QOI = "Q3",
                            target1 = "Immune target", target2 = "Tumor target")
# example 4: positioning of labels Q4 (samples low in both targets)
direction_and_axis_quadrants(df = df_harmonized, QOI = "Q4",
                            target1 = "Immune target", target2 = "Tumor target")

}
