% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/89_general_helpers.R
\name{error_handling}
\alias{error_handling}
\title{Error handlers}
\usage{
error_handling(object = NULL, metric = NULL, gofmetric = NULL, progress = NULL)
}
\description{
Error handlers
}
\keyword{internal}
