% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_hp_helpers.R
\name{fit_and_gof}
\alias{fit_and_gof}
\title{Fit all possible submodels and obtain goodnesses of fit}
\usage{
fit_and_gof(depvar, expl_df, fam, ncores, progress, gofmetric, class)
}
\description{
Fit all possible submodels and obtain goodnesses of fit
}
\keyword{internal}
