% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relweights_helpers.R
\name{fit_null_model}
\alias{fit_null_model}
\alias{fit_null_model.glm}
\alias{fit_null_model.gam}
\alias{fit_null_model.gamlss}
\title{Fit a null model (model without any explanatory variables)}
\usage{
fit_null_model(object, ...)

\method{fit_null_model}{glm}(object, ...)

\method{fit_null_model}{gam}(object, ...)

\method{fit_null_model}{gamlss}(object, ...)
}
\arguments{
\item{object}{Model object, one of the ones supported in \link{scam}}

\item{...}{Other arguments}
}
\value{
a fitted model object
}
\description{
Fit a null model (model without any explanatory variables)
}
