% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correlation.R
\name{get_correlation}
\alias{get_correlation}
\title{Calculate correlation between two genes}
\usage{
get_correlation(df, target1, target2, cor_method, grouping_var = NA)
}
\arguments{
\item{df}{dataframe with expression values for each indication (as obtained from Profiler)}

\item{target1}{Official gene symbol of gene 1}

\item{target2}{Official gene symbol of gene 2}

\item{cor_method}{Correlation method used}

\item{grouping_var}{groups object based on this variable; intr if not provided}
}
\value{
tibble with correlation value
}
\description{
Calculate correlation between two genes
}
\examples{
# prepare data
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))


# example 1: pearson correlation between target1 and target2 per intr group
get_correlation(df = df_harmonized, target1 = "Immune target", target2 = "Tumor target",
cor_method = "pearson")

# example 2: spearman correlation between target1 and target2 per indication group
get_correlation(df = df_harmonized, target1 = "Immune target", target2 = "Tumor target",
               cor_method = "spearman", grouping_var = "indication")

}
