% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_table.R
\name{get_expr_table}
\alias{get_expr_table}
\title{create table with sample numbers and median expression per intr group (indication + treatment)}
\usage{
get_expr_table(df, method = c("mean", "median", "Q3"), grouping_var = NA)
}
\arguments{
\item{df}{dataframe in long format containing intr (indication + treatment), value and symbol columns}

\item{method}{the summary stat method to use, either "mean" or "median"}

\item{grouping_var}{groups object based on this variable; intr if not provided}
}
\value{
sample number, median expression and median thresholds of all available genes per intr group
}
\description{
create table with sample numbers and median expression per intr group (indication + treatment)
}
\examples{
# prepare data
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))


# example1: median expression of each intr group and all genes in df_harmonized
get_expr_table(df = df_harmonized, method = "median")

# example2: mean expression of each indication group and all genes in df_harmonized
get_expr_table(df = df_harmonized, method = "mean", grouping_var = "indication")

}
