% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numbers.R
\name{get_numbers}
\alias{get_numbers}
\title{Function to calculate the number of samples per group in the data set}
\usage{
get_numbers(df, grouping_var = NA)
}
\arguments{
\item{df}{object with expression data frame for each indication}

\item{grouping_var}{groups object based on this variable; intr if not provided}
}
\value{
The number of samples per indication
}
\description{
Function to calculate the number of samples per group in the data set
}
\examples{
# prepare data
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))


# example1: sample numbers per intr group
get_numbers(df = df_harmonized)

# example2: sample numbers per indiation group
get_numbers(df = df_harmonized, grouping_var = "indication")
}
