% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perc_gene.R
\name{get_perc_gene}
\alias{get_perc_gene}
\title{Function to calculate the percentage of samples above the threshold expression per group in the data set}
\usage{
get_perc_gene(df, gene, method = c("median", "mean", "Q3"), grouping_var = NA)
}
\arguments{
\item{df}{object with expression data frame for each indication}

\item{gene}{official gene symbol of the gene}

\item{method}{"median" or "mean"; The method used for calculating the threshold value.}

\item{grouping_var}{groups object based on this variable; intr if not provided}
}
\value{
The percentage of samples above the threshold
}
\description{
Function to calculate the percentage of samples above the threshold expression per group in the data set
}
\examples{
# prepare data
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))


# example1: percentage of samples above median threshold per intr group
get_perc_gene(df = df_harmonized, gene = "Immune target",
             method = "median")

# example2: percentage of samples above mean threshold per indication group
get_perc_gene(df = df_harmonized, gene = "Immune target",
             method = "mean", grouping_var = "indication")
}
