% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perc_pop.R
\name{get_perc_pop}
\alias{get_perc_pop}
\title{Returns a dataframe with percentage of samples per quadrant or Quadrant of interest (QOI)
Q1: Low target1; High target 2
Q2: High target1; High target 2
Q3: Low target1; Low target 2
Q4: High target1; Low target 2}
\usage{
get_perc_pop(
  df,
  target1,
  target2,
  threshold_method = c("median", "mean", "Q3"),
  QOI = c("Q1", "Q2", "Q3", "Q4", "all"),
  grouping_var = NA
)
}
\arguments{
\item{df}{the dataframe}

\item{target1}{Official gene symbol of gene 1}

\item{target2}{Official gene symbol of gene 2}

\item{threshold_method}{Threshold method to use; median or mean or Q3 (75th quantile)}

\item{QOI}{Quadrant of interest among 4 quadrants}

\item{grouping_var}{groups object based on this variable; intr if not provided}
}
\value{
the dataframe with percentage of samples per quadrant or Quadrant of interest (QOI)
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        #
Q1      #       Q2
}\if{html}{\out{</div>}}

T            #
a            #
r            #
g ########################
e            #
t            #
2            #
Q3      #       Q4
#
target 1
}
\examples{
# prepare data
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))


# example1: percentage of samples with high expression of both targets (Q2)
get_perc_pop(df = df_harmonized,
            target1 = "Immune target",
            target2 = "Tumor target",
            threshold_method = "median", QOI = "Q2")

# example2: percentage of samples for all quadrants of interest (QOI) per indication group
get_perc_pop(df = df_harmonized,
            target1 = "Immune target",
            target2 = "Tumor target",
            threshold_method = "median", QOI = "all",
            grouping_var = "indication")
}
