% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_threshold.R
\name{get_threshold}
\alias{get_threshold}
\title{Function to calculate the threshold expression values for a specified gene
By calculating the median, mean or Q3 expression over all samples in the data set}
\usage{
get_threshold(df, gene, method = c("median", "mean", "Q3"))
}
\arguments{
\item{df}{object with expression data frame as harmonized by @harmonized_df}

\item{gene}{Official gene symbol}

\item{method}{"median", "mean" or "Q3"; The method used for calculating the threshold value}
}
\value{
The threshold expression value for the specified gene
}
\description{
Function to calculate the threshold expression values for a specified gene
By calculating the median, mean or Q3 expression over all samples in the data set
}
\examples{
# prepare data
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))


# example1: median threshold
get_threshold(df = df_harmonized, gene = "Immune target", method = "median")

# example2: mean threshold
get_threshold(df = df_harmonized, gene = "Immune target", method = "mean")
}
