% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_gof.R
\name{gof}
\alias{gof}
\title{Function to obtain goodness-of-fit from regression object}
\usage{
gof(object, ...)
}
\arguments{
\item{object}{Regression object. One of \code{vibe:::supported_classes}.}

\item{...}{Additional arguments}
}
\description{
Function to obtain goodness-of-fit from regression object
}
