% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_df.R
\name{harmonize_df}
\alias{harmonize_df}
\title{harmonize dataframe for VIBE}
\usage{
harmonize_df(
  df,
  col_patientid = NA,
  col_sampleid = NA,
  col_indication,
  col_treatment = NA,
  col_therapy = NA,
  col_symbol,
  col_value,
  unit,
  keep_extra_cols = NA
)
}
\arguments{
\item{df}{dataframe with expression values for each indication in long format}

\item{col_patientid}{column name for column that contains unique patient identifiers (optional column)}

\item{col_sampleid}{column name for colun that contains unique sample identifiers  (optional column)}

\item{col_indication}{column name for column that contains indications}

\item{col_treatment}{column name for column that contains indications (pre/post-treatment)  (optional column)}

\item{col_therapy}{column name for column that contains therapy (what type of treatment, e.g. chemo)  (optional column)}

\item{col_symbol}{column name for column that contains gene symbols}

\item{col_value}{column name for column that contains normalized count values}

\item{unit}{unit that is used for the normalized counts (e.g. tpm or cpm)}

\item{keep_extra_cols}{vector with column names if there are additional columns you wish to keep}
}
\value{
df dataframe in format that is appropriate for VIBE functions
}
\description{
harmonize dataframe for VIBE
}
\examples{
df_harmonized <- harmonize_df(df = VIBE_data,
col_patientid = "patient_no",
col_sampleid = "analysis",
col_indication = "tumor",
col_treatment = "treatment_flag",
col_symbol = "gene",
col_value = "log2_tpm",
unit = "tpm")
}
