% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_df_pathway.R
\name{harmonize_df_pathway}
\alias{harmonize_df_pathway}
\title{Calculate average expression of gene list combined for pathways}
\usage{
harmonize_df_pathway(df, gene_panel_list, grouping_cols = NA)
}
\arguments{
\item{df}{dataframe with expression values for each indication (as returned by harmonize_df function)}

\item{gene_panel_list}{Official gene symbol of genes compiled in list}

\item{grouping_cols}{groups object based on these columns; if default columns not provided}
}
\value{
dataframe with expression values for each gene/pathway and indication in long format
}
\description{
Calculate average expression of gene list combined for pathways
}
\examples{
# Creating lists ----------------------------------------------------------------------
pathway_1 <- c("Gene 14","Gene 30","Gene 20","Gene 22","Gene 29",
"Gene 23","Gene 1","Gene 32","Gene 33")
pathway_2 <- c("Gene 21","Gene 11","Immune target","Gene 18","Gene 8","Gene 16")
pathway_3 <- c("Gene 2" ,"Gene 3" ,"Gene 5" ,"Gene 6","Gene 7","Gene 10","Gene 15",
"Gene 17","Gene 18","Gene 24","Gene 25","Gene 28","Gene 31","Gene 35")
pathway_4 <- c('Gene 19',"Gene 12","Gene 26")
immune_target <- "Immune target"
tumor_target <- "Tumor target"
gene_panel_list <- list( pathway_1,pathway_2,pathway_3,pathway_4, immune_target, tumor_target)
names(gene_panel_list) <- c("pathway_1", "pathway_2","pathway_3","pathway_4",
"immune_target","tumor_target")

# Calling function ----------------------------------------------------------------------

df_harmonized <- harmonize_df(df = VIBE_data,
col_patientid = "patient_no",
col_sampleid = "analysis",
col_indication = "tumor",
col_treatment = "treatment_flag",
col_symbol = "gene",
col_value = "log2_tpm",
unit = "tpm")
df_avg_pathways <- harmonize_df_pathway(df_harmonized, gene_panel_list)
}
