% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_sample_expression.R
\name{heatmap_sample_expression}
\alias{heatmap_sample_expression}
\title{Heatmap representing the median expression of indications}
\usage{
heatmap_sample_expression(
  df,
  method = "median",
  grouping_var = NA,
  cluster_switch = NULL,
  scale = TRUE,
  cluster_columns = TRUE,
  column_title = NA,
  plot_groups = NA,
  order_4_heatmap = NA
)
}
\arguments{
\item{df}{dataframe in long format containing intr (indication + treatment), value and symbol columns}

\item{method}{the summary stat method to use, either "mean" or "median", median is standard}

\item{grouping_var}{groups object based on this variable; intr if not provided}

\item{cluster_switch}{if you would like to cluster your rows based on pathways etc. default is NULL}

\item{scale}{scale the expression of genes in range 0-1 (default is TRUE)}

\item{cluster_columns}{if the heatmap grouping variables are to be clustered (default is TRUE)}

\item{column_title}{provide a text for title of your heatmap (default is provided)}

\item{plot_groups}{Character vector with values you wish to include in the plot, threshold will be calculated
for target1 and target2 on all data in df. plot_groups values must be present in the grouping_var column.}

\item{order_4_heatmap}{Character vector of the order of genes in heatmap (helpful for reassigning rowsplit)}
}
\value{
heatmap with cells colored based on the percentage of samples darkblue the lowest and magenta the highest
}
\description{
Heatmap representing the median expression of indications
}
\examples{
# prepare data
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))

# example1: Heatmap of expression data for each comparison for default grouping var "intr"
heatmap_sample_expression(df_harmonized, method = "median")

# example2: Heatmap of expression data for each comparison for grouping var "indication"
heatmap_sample_expression(df_harmonized, method = "mean",  grouping_var="indication")

# example3: Splitting the heatmap in groups

rowsplit <-rep(c("Group1", "Group2","Group3", "Group4","IT","TT"), c(9,8,9,7,1,1))
heatmap_sample_expression(df_harmonized, method="median", cluster_switch=rowsplit, plot_groups = c("NSCLC","RCC"))

# example4: Reordering the heatmap and redefining the groups based on user-defined pathways

reorder_list <- c("Gene 14","Gene 30","Gene 20","Gene 22","Gene 29","Gene 23","Gene 1","Gene 32","Gene 33","Gene 21","Gene 11","Gene 18","Gene 8","Gene 16","Gene 2" ,"Gene 3" ,"Gene 5" ,"Gene 6","Gene 7","Gene 10","Gene 15","Gene 17","Gene 24","Gene 25","Gene 28","Gene 31","Gene 35",'Gene 19',"Gene 12","Gene 26","Gene 4","Gene 9","Immune target","Tumor target")

rowsplit_new <-rep(c("Pathway1", "Pathway2","Pathway3", "Pathway4","G4","G9","IT","TT"), c(9,5,13,3,1,1,1,1))
heatmap_sample_expression(df_harmonized, method="median", cluster_switch=rowsplit_new, order_4_heatmap = reorder_list)



}
