% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_samples_above_median.R
\name{heatmap_samples_above_median}
\alias{heatmap_samples_above_median}
\title{Heatmap representing the percentage of samples above median expression}
\usage{
heatmap_samples_above_median(
  df,
  method = c("median", "mean"),
  grouping_var = NA,
  cluster_switch = NULL,
  cluster_columns = TRUE,
  column_title = NA,
  plot_groups = NA,
  order_4_heatmap = NA
)
}
\arguments{
\item{df}{dataframe in long format containing intr (indication + treatment), value and symbol columns}

\item{method}{the summary stat method to use, either "mean" or "median"}

\item{grouping_var}{groups object based on this variable; intr if not provided}

\item{cluster_switch}{if you would like to cluster your rows based on pathways etc. default is NULL}

\item{cluster_columns}{if the heatmap grouping variables are to be clustered (default is TRUE)}

\item{column_title}{provide a text for title of your heatmap (default is provided)}

\item{plot_groups}{Character vector which groups from column defined at grouping_var you wish to include in the plot. Median expression will be calculated from all groups in df.}

\item{order_4_heatmap}{Character vector of the order of genes in heatmap (helpful for reassigning rowsplit)}
}
\value{
heatmap with cells colored based on the percentage of samples darkblue the lowest and magenta the highest
}
\description{
Heatmap representing the percentage of samples above median expression
}
