% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/89_general_helpers.R
\name{is_any_multiple_classes}
\alias{is_any_multiple_classes}
\title{Multiple class checker}
\usage{
is_any_multiple_classes(object, classes)
}
\description{
Multiple class checker
}
\keyword{internal}
