% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{obtain_gof}
\alias{obtain_gof}
\title{Obtain goodness-of-fit figure from regression object}
\usage{
obtain_gof(object, ...)
}
\arguments{
\item{object}{Regression object. One of \code{vibe:::supported_classes}.}

\item{...}{Additional arguments}
}
\description{
blabla
}
