% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{obtain_gof_r2_mcfadden}
\alias{obtain_gof_r2_mcfadden}
\title{Obtain goodness of fit figure for McFadden's Pseudo R2}
\usage{
obtain_gof_r2_mcfadden(object, m0)
}
\arguments{
\item{object}{Regression object. One of the supported classes documented in
\link{scam}.}

\item{m0}{This is the "empty model", the model without any explanatory variables.}
}
\description{
Obtain goodness of fit figure for McFadden's Pseudo R2
}
\references{
D. McFadden. Conditional logit analysis of qualitative choice behavior. In P. Zarembka, editor, Frontiers in Econometrics, chapter Four, pages 104–142. Academic Press, New York, 1974.
}
\keyword{internal}
