% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{obtain_gof_r2e}
\alias{obtain_gof_r2e}
\title{Obtain goodness of fit figure for Estrella's Pseudo R2}
\usage{
obtain_gof_r2e(object, m0, depvar)
}
\arguments{
\item{object}{Regression object. One of the supported classes documented in
\link{scam}.}

\item{m0}{This is the "empty model", the model without any explanatory variables.}
}
\description{
Obtain goodness of fit figure for Estrella's Pseudo R2
}
\references{
Arturo Estrella (1998) A New Measure of Fit for Equations With Dichotomous Dependent Variables, Journal of Business & Economic Statistics, 16:2, 198-205, DOI: 10.1080/07350015.1998.10524753
}
\keyword{internal}
