% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/89_general_helpers.R
\name{pcapply}
\alias{pcapply}
\title{Parallel Column Apply}
\usage{
pcapply(X, FUN, ncores, progress = TRUE)
}
\description{
works like apply but also coerces into nice format like sapply. Gives out a single vector of numbers
}
\keyword{internal}
