% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box_pre_post.R
\name{plot_box_pre_post}
\alias{plot_box_pre_post}
\title{Plot a boxplot pre and post treatment per gene per indication with Kruskal Wallis test, Fold change and sample counts}
\usage{
plot_box_pre_post(
  df,
  dataset = NA,
  symbol_list,
  method = c("mean", "median"),
  plot_groups = NA,
  treatment_levels = c("pre-treatment", "post-treatment")
)
}
\arguments{
\item{df}{dataframe harmonized by VIBE's harmonize_df function with paired pre/post info}

\item{dataset}{name of dataset to be mentioned for captioning and sorting plot-axis.}

\item{symbol_list}{list of genes for creating the plot}

\item{method}{the summary stat method to use, either "mean" or "median"}

\item{plot_groups}{Character vector with values you wish to include in the plot, threshold will be calculated for target1 and target2 on all data in df. plot_groups values must be present in the grouping_var column.}

\item{treatment_levels}{pre and post-treatment levels}
}
\value{
boxplot comparing the statistical differences of the sample distribution per gene, per indication
}
\description{
Plot a boxplot pre and post treatment per gene per indication with Kruskal Wallis test, Fold change and sample counts
}
\examples{
# prepare data
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))


# example1:
plot_box_pre_post(df = df_harmonized,
                 dataset = "VIBE dummy data (not real data!)",
                 symbol_list = c("Immune target", "Tumor target", "Gene 1"),
                 method = "median")
# example2: limit indications plotted to selection
plot_box_pre_post(df = df_harmonized,
                 dataset = "VIBE dummy data (not real data!)",
                 symbol_list = c("Immune target", "Tumor target", "Gene 1"),
                 method = "median", plot_groups = c("CERV", "NSCLC", "SARC"))
# example3: calculate fold change based on mean values instead of median
plot_box_pre_post(df = df_harmonized,
                 dataset = "VIBE dummy data (not real data!)",
                 symbol_list = c("Immune target", "Tumor target", "Gene 1"),
                 method = "mean", plot_groups = c("CERV", "NSCLC", "SARC"))
}
