% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression_box.R
\name{plot_expression_box}
\alias{plot_expression_box}
\title{Function to make boxplot of median/mean expression of a gene per indication}
\usage{
plot_expression_box(
  df,
  gene,
  method = c("median", "mean"),
  order = c("T", "F"),
  dataset = NA,
  plot_groups = NA,
  grouping_var = NA
)
}
\arguments{
\item{df}{VIBE harmonized dataframe}

\item{gene}{Official gene symbol}

\item{method}{"median" or "mean"; method to calculate the summarized expression for the gene}

\item{order}{"T" or "F"; ordering x-axis of the plot in decreasing order}

\item{dataset}{character string with name of dataset, used in caption}

\item{plot_groups}{Character vector with values you wish to include in the plot, however threshold will be calculated for target1 and target2 on all data in df. plot_groups values must be present in the grouping_var column.}

\item{grouping_var}{groups object based on this variable; intr if not provided}
}
\value{
ggplot barplot with expression median/mean for the gene
}
\description{
Function to make boxplot of median/mean expression of a gene per indication
}
\examples{
# prepare data
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))

# example1: plot box for each unique intr value
plot_expression_box(df = df_harmonized, gene = "Tumor target", method = "median", order = "F",
                   dataset = "VIBE dummy data (not real data!)")
# example2: order boxes based on percentage above threshold
plot_expression_box(df = df_harmonized, gene = "Tumor target", method = "median", order = "T",
                   dataset = "VIBE dummy data (not real data!)")
# example3: group data by indication & plot selection of indications
plot_expression_box(df = df_harmonized, gene = "Tumor target", method = "median", order = "T",
dataset = "VIBE dummy data (not real data!)",
grouping_var = "indication", plot_groups = c("RCC", "NSCLC", "PDAC"))
}
