% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression_box_gtex_tcga.R
\name{plot_expression_box_gtex_tcga}
\alias{plot_expression_box_gtex_tcga}
\title{Plot a boxplot comparing GTEX and TCGA samples for selected gene with Percentage of samples above median, Kruskal Wallis test, Fold change and sample counts}
\usage{
plot_expression_box_gtex_tcga(
  df,
  gene,
  threshold_method,
  dataset,
  plot_primary_site = NA
)
}
\arguments{
\item{df}{dataframe harmonized by VIBE's harmonize_df function, with the indication column being a concatenation of the study and primary_site columns of the XENA dataset.}

\item{gene}{name of genes for creating the plot}

\item{threshold_method}{the summary stat method to use, either "mean" or "median"}

\item{dataset}{character string with name of dataset, used in caption}

\item{plot_primary_site}{Character vector with primary_site values (name of tissue column in Xena dataset) you wish to include in the plot, however threshold will be calculated for target1 and target2 on all data in df.}
}
\value{
boxplot comparing the statistical differences of the sample distribution per gene in GTEX vs TCGA
}
\description{
Plot a boxplot comparing GTEX and TCGA samples for selected gene with Percentage of samples above median, Kruskal Wallis test, Fold change and sample counts
}
