% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression_box_split.R
\name{plot_expression_box_split}
\alias{plot_expression_box_split}
\title{Plot a boxplot comparing samples from multiple groups (unmatched) for selected gene with Percentage of samples above median, Kruskal Wallis test, Fold change and sample counts}
\usage{
plot_expression_box_split(
  df,
  gene,
  threshold_method,
  dataset = "Dataset",
  grouping_var = "intr",
  plot_groups = NA,
  splitting_var,
  reverse = FALSE
)
}
\arguments{
\item{df}{dataframe harmonized by VIBE's harmonize_df function, with the indication column being a concatenation of the study and primary_site columns of the XENA dataset.}

\item{gene}{name of genes for creating the plot}

\item{threshold_method}{the summary stat method to use, either "mean" or "median"}

\item{dataset}{character string with name of dataset, used in caption}

\item{grouping_var}{column name which you want to group data by (intr if left to NA)}

\item{plot_groups}{Character vector with values you wish to include in the plot, threshold will be calculated for target1 and target2 on all data in df. plot_groups values must be present in the grouping_var column.}

\item{splitting_var}{column name of column that should be used to color plot boxplots with.}

\item{reverse}{TRUE or FALSE, if FALSE fold chanae is calculated as b/a, if TRUE as a/b}
}
\value{
boxplot comparing the statistical differences of the sample distribution per gene in GTEX vs TCGA
}
\description{
Plot a boxplot comparing samples from multiple groups (unmatched) for selected gene with Percentage of samples above median, Kruskal Wallis test, Fold change and sample counts
}
\examples{
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))


# example1: calculate fold change between two groups and show as boxplots
plot_expression_box_split(df = df_harmonized,
                         gene = "Tumor target",
                         threshold_method = "median",
                         dataset = "VIBE dummy data (not real data!)",
                         splitting_var = "database")

# example2: group x-axis by indication instead of intr
plot_expression_box_split(df = df_harmonized,
                         gene = "Tumor target",
                         threshold_method = "median",
                         dataset = "VIBE dummy data (not real data!)",
                         splitting_var = "database",
                         grouping_var = "indication")

# example3: limit number of groups plotted on x-axis and reverse numerator/denominator
# as the output is ggplot2 based, you can use additional ggplot2 functionality
plot_expression_box_split(df = df_harmonized,
                         gene = "Tumor target",
                         threshold_method = "median",
                         dataset = "VIBE dummy data (not real data!)",
                         splitting_var = "database",
                         grouping_var = "indication",
                         plot_groups = c("BLCA", "CRC", "TNBC"),
                         reverse = TRUE) +
 ggplot2::expand_limits(y = 0)

}
