% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression_scatterplot.R
\name{plot_expression_scatter}
\alias{plot_expression_scatter}
\title{Multi-panel plot representing scatterplot per quadrant between two different genes.}
\usage{
plot_expression_scatter(
  df,
  target1,
  target2,
  target3 = NA,
  QOI = c("Q1", "Q2", "Q3", "Q4"),
  cor_method = c("spearman", "pearson"),
  threshold_method = c("median", "mean", "Q3"),
  dataset = NA,
  plot_groups = NA,
  grouping_var = NA,
  facet_ncol = 4
)
}
\arguments{
\item{df}{dataframe with expression values for each indication (as obtained from Profiler)}

\item{target1}{Official gene symbol of gene 1}

\item{target2}{Official gene symbol of gene 2}

\item{target3}{Official gene symbol of gene 3. Will be used to color dots by. If left NA dots will be colored by quadrant.}

\item{QOI}{Which of the four quadrants are of interest for the question; Q1 (Top-left; high-low); Q2 (Top-right; high-high); Q3 (bottom-left; low-high) and Q4 (bottom right; low-low)}

\item{cor_method}{Correlation method used}

\item{threshold_method}{Threshold method to use; median or mean or Q3 (75th quantile)}

\item{dataset}{dataset to be mentioned for captioning}

\item{plot_groups}{Character vector with values you wish to include in the plot, threshold will be calculated for target1 and target2 on all data in df. plot_groups values must be present in the grouping_var column.}

\item{grouping_var}{groups object based on this variable; intr if not provided}

\item{facet_ncol}{number of columns to be plotted}
}
\value{
ggplot2 scatterplot with information of the percentage of samples in the selected quadrant by selected correlation method; barplot with correlation values and the table with correlation values between two genes.
}
\description{
Multi-panel plot representing scatterplot per quadrant between two different genes.
}
\examples{
# prepare data
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))


# example1:
plot_expression_scatter(df = df_harmonized, target1 = "Immune target", target2 = "Tumor target",
                       QOI = "Q2", cor_method = "spearman", threshold_method = "median",
                       dataset = "VIBE dummy data (not real data!)")
# example2: change grouping_var to indication
plot_expression_scatter(df = df_harmonized, target1 = "Immune target", target2 = "Tumor target",
                       QOI = "Q2", cor_method = "spearman", threshold_method = "median",
                       dataset = "VIBE dummy data (not real data!)",
                       grouping_var = "indication")
# example3: limit plotted indications to selection,
# also color datapoints by expression of a third gene
plot_expression_scatter(df = df_harmonized, target1 = "Immune target", target2 = "Tumor target",
                       QOI = "Q2", cor_method = "spearman", threshold_method = "median",
                       dataset = "VIBE dummy data (not real data!)",
                       grouping_var = "indication", plot_groups = c("BRCA", "NSCLC", "OV", "STAD"),
                       target3 = "Gene 9",
                       facet_ncol = 2)

}
