% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perc_pop.R
\name{plot_perc_pop}
\alias{plot_perc_pop}
\title{Plot percentage of samples per quadrant or Quadrant of interest (QOI)}
\usage{
plot_perc_pop(
  df,
  target1,
  target2,
  threshold_method = c("median", "mean", "Q3"),
  QOI = c("Q1", "Q2", "Q3", "Q4", "all"),
  order = c("T", "F"),
  dataset = NA,
  grouping_var = NA,
  plot_groups = NA
)
}
\arguments{
\item{df}{the dataframe}

\item{target1}{Official gene symbol of gene 1}

\item{target2}{Official gene symbol of gene 2}

\item{threshold_method}{Threshold method to use; median or mean or Q3 (75th quantile)}

\item{QOI}{Quadrant of interest among 4 quadrants}

\item{order}{True or False to sort the a axis in decreasing order ofr the quadrant of interest (Will not work for all quadrants)}

\item{dataset}{Name of your dataset for captioning}

\item{grouping_var}{column name which you want to group data by (intr if left to NA)}

\item{plot_groups}{Character vector with values you wish to include in the plot, threshold will be calculated for target1 and target2 on all data in df. plot_groups values must be present in the grouping_var column.}
}
\value{
the dataframe with percentage of samples per quadrant or Quadrant of interest (QOI)
}
\description{
Plot percentage of samples per quadrant or Quadrant of interest (QOI)
}
\examples{
# prepare data
data("VIBE_data")
df_harmonized = harmonize_df(df = VIBE_data,
                            col_patientid = "patient_no",
                          col_sampleid = "analysis",
                          col_indication = "tumor",
                          col_treatment = "treatment_flag",
                          col_symbol = "gene",
                          col_value = "log2_tpm",
                          unit = "tpm",
                          keep_extra_cols = c("database"))


# example1: percentage of samples in Q1 (low target1 and high target 2)
plot_perc_pop(df = df_harmonized, target1 = "Immune target", target2 = "Tumor target",
             threshold_method = "median", QOI = "Q1",
             dataset = "VIBE dummy data", order = "F")
# example2: percentage of samples in Q1 (low target1 and high target 2) sorted
plot_perc_pop(df = df_harmonized, target1 = "Immune target", target2 = "Tumor target",
             threshold_method = "median",
             QOI = "Q1", dataset = "VIBE dummy data", order = "T")
# example3: percentage of all quadrants per indication
plot_perc_pop(df = df_harmonized, target1 = "Immune target", target2 = "Tumor target",
             threshold_method = "median",
             QOI = "all", dataset = "VIBE dummy data", order = "T",
             grouping_var = "indication", plot_groups = c("CRC", "PRAD", "UTEN"))
}
