% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_output.R
\name{print.vibe}
\alias{print.vibe}
\title{Print function for 'vibe' objects}
\usage{
\method{print}{vibe}(x, ...)
}
\arguments{
\item{x}{Any object with class 'vibe'}

\item{...}{Additional arguments}
}
\description{
Print function for 'vibe' objects
}
