% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_downloadable_df.R
\name{print_downloadable_df}
\alias{print_downloadable_df}
\title{For HTML notebooks, print a downloadable table that also allows for searching}
\usage{
print_downloadable_df(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a downloadable table that is nicely printed in a html notebook
}
\description{
For HTML notebooks, print a downloadable table that also allows for searching
}
\examples{
# example: print any dataframe to the viewer (or nicely printed in a notebook)
data("VIBE_data")
print_downloadable_df(VIBE_data[1:10,])
}
