% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_relweights.R
\name{rel_weights}
\alias{rel_weights}
\title{Compute Relative Weights}
\usage{
rel_weights(expl_df, depvar, fam, class, gofmetric, param = "mu")
}
\description{
Compute Relative Weights
}
\keyword{internal}
