% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_location_per_quadrant.R
\name{sample_location_per_quadrant}
\alias{sample_location_per_quadrant}
\title{Obtain the quadrant specific direction of selected genes and returns positioning of labels on the quadrant}
\usage{
sample_location_per_quadrant(x, y, xintercept, yintercept)
}
\arguments{
\item{x}{the expression values of target1}

\item{y}{the expression values of target2}

\item{xintercept}{threshold for target 1 as per the selected threshold method}

\item{yintercept}{threshold for target 2 as per the selected threshold method}
}
\value{
the quadrant information per sample based on the selected targets to be added to the quad_summary table
}
\description{
Obtain the quadrant specific direction of selected genes and returns positioning of labels on the quadrant
}
\note{
@add_quadrant_info acts as a wrapper around this function
}
