% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sat}
\alias{sat}
\title{Customer Satisfaction in North Macedonia}
\format{
An object of class \code{data.frame} with 452 rows and 17 columns.
}
\source{
\url{https://www.kaggle.com/vdimitrievska/patient-satisfaction-dataset}
}
\usage{
sat
}
\description{
Customer Satisfaction in North Macedonia
}
\keyword{datasets}
