% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scam}
\alias{scam}
\title{Supported Classes and Metrics (SCAM)}
\format{
An object of class \code{data.frame} with 24 rows and 4 columns.
}
\usage{
scam
}
\description{
A \code{data.frame} object with 4 columns:
\describe{
\item{supported_classes}{R object classes}
\item{varimp_measure}{Variable importance measures}
\item{gof_metric}{Goodness-of-fit metrics}
\item{implemented}{A boolean value depicting whether this specific
combination of R class, variable importance measure and
goodness of fit metric is implemented in \code{vibe}.}
}
}
\keyword{datasets}
