% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{summary.vibe}
\alias{summary.vibe}
\title{Summary function for 'vibe' objects}
\usage{
\method{summary}{vibe}(object, ...)
}
\arguments{
\item{object}{Any object with class 'vibe'}

\item{...}{Additional arguments}
}
\description{
Summary function for 'vibe' objects
}
