% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibe-package.R
\docType{package}
\name{vibe-package}
\alias{vibe-package}
\alias{_PACKAGE}
\alias{testduplicate-package}
\title{vibe: Variable Importance Beyond Linear Models}
\description{
Functions for calculating variable importance metrics for glm, gam and gamlss classes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Stan125/vibe}
  \item Report bugs at \url{https://github.com/Stan125/vibe/issues}
}

}
\author{
\strong{Maintainer}: Stanislaus Stadlmann \email{stanislaus@stadlmann.cm} (\href{https://orcid.org/0000-0001-6542-6342}{ORCID})

}
\keyword{internal}
