% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_gam_vibes.R
\name{vibe.gam}
\alias{vibe.gam}
\title{Variable Importance calculation for a \code{gam} object}
\usage{
\method{vibe}{gam}(
  object,
  metric = "hp",
  gofmetric = "R2e",
  ncores = 1,
  progress = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{gam} object, typically result of \link[mgcv:gam]{mgcv::gam}}

\item{...}{}
}
\description{
Variable Importance calculation for a \code{gam} object
}
