% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_vibes.R
\name{vibe.gamlss}
\alias{vibe.gamlss}
\title{Variable Importance calculation for a \code{gamlss} object}
\usage{
\method{vibe}{gamlss}(object, varimp = "hp", gof = "R2e", ncores = 1, progress = TRUE, ...)
}
\arguments{
\item{object}{A \code{gam} object, typically result of \code{\link[mgcv:gam]{mgcv::gam()}}.}

\item{varimp}{One of \code{c("hp", "relweights")}, which stand for hierarchical
partitioning and relative weights.}

\item{gof}{Goodness-of-fit metric, the changes of which shall be
analysed}

\item{ncores}{Number of cores used for the model fitting process, happening
in \code{\link{part_core}}.}

\item{progress}{Boolean. Do you want to see a progress bar?}

\item{...}{Other arguments}
}
\description{
\code{vibe.gamlss} takes a fitted \code{\link[gamlss:gamlss]{gamlss::gamlss()}} object and calculates
variable importance metrics by fitting the submodels required, extracting the
desired goodness-of-fit metric and applying variable importance metrics to
it.
}
