% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_glm_vibes.R
\name{vibe.glm}
\alias{vibe.glm}
\title{Variable Importance calculation for a \code{glm} object}
\usage{
\method{vibe}{glm}(object, metric = "hp", gofmetric = "R2e", ncores = 1, progress = TRUE)
}
\arguments{
\item{object}{A \code{glm} object, typically result of \code{\link[=glm]{glm()}}.}

\item{metric}{One of \code{c("hp", "relweights")}, which stand for hierarchical
partitioning and relative weights}

\item{gofmetric}{Goodness-of-fit metric, the changes of which shall be
analysed}

\item{ncores}{Number of cores used for the model fitting process, happening
in \code{\link{part_core}}.}

\item{progress}{Boolean. Do you want to see a progress bar?}
}
\description{
\code{vibe.glm} takes a fitted \code{\link[=glm]{glm()}} object and calculates
variable importance metrics by fitting the submodels required, extracting the
desired goodness-of-fit metric and applying variable importance metrics to
it.
}
