% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_vibes.R
\name{vibe.lm}
\alias{vibe.lm}
\title{Variable Importance calculation for an \code{lm} object}
\usage{
\method{vibe}{lm}(object, varimp = "hp", gof = "R2e", ncores = 1, progress = TRUE, ...)
}
\arguments{
\item{object}{A \code{lm} object, typically result of \code{\link[=lm]{lm()}}.}

\item{varimp}{One of \code{c("hp", "relweights")}, which stand for hierarchical
partitioning and relative weights.}

\item{gof}{Goodness-of-fit metric, the changes of which shall be
analysed}

\item{ncores}{Number of cores used for the model fitting process, happening
in \code{\link{part_core}}.}

\item{progress}{Boolean. Do you want to see a progress bar?}

\item{...}{Other arguments}
}
\description{
\code{vibe.lm} takes a fitted \code{\link[=lm]{lm()}} object and calculates
variable importance metrics by fitting the submodels required, extracting the
desired goodness-of-fit metric and applying variable importance metrics to
it.
}
\examples{
iris_lm <- lm(Sepal.Width ~ ., data = iris)
vibe(iris_lm, gof = "R2Mac", varimp = "hp")

}
