% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_helpers.R
\name{fit_and_gof}
\alias{fit_and_gof}
\title{Fit all possible submodels and obtain goodnesses of fit}
\usage{
fit_and_gof(
  depvar,
  expl_df,
  fam,
  ncores,
  progress,
  gof,
  class,
  depvar_name,
  base_df,
  param,
  object
)
}
\description{
Fit all possible submodels and obtain goodnesses of fit
}
\keyword{internal}
