% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{make_vibe}
\alias{make_vibe}
\title{Coerce results from algorithms into vibes object}
\usage{
make_vibe(results, varimp, depvar_name, class)
}
\description{
Coerce results from algorithms into vibes object
}
\keyword{internal}
