% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{pcapply}
\alias{pcapply}
\title{Parallel Column Apply}
\usage{
pcapply(x_mat, fun, ncores, progress = TRUE)
}
\description{
works like apply but also coerces into nice format like sapply.
Gives out a single vector of numbers
}
\keyword{internal}
