% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data.R
\docType{data}
\name{sat}
\alias{sat}
\title{Customer Satisfaction in North Macedonia}
\format{
An object of class \code{data.frame} with 453 rows and 17 columns.
}
\usage{
sat
}
\description{
\link{https://www.kaggle.com/vdimitrievska/patient-satisfaction-dataset}
}
\keyword{datasets}
