% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_vibes.R
\name{vibe.gam}
\alias{vibe.gam}
\title{Variable Importance calculation for a \code{gam} object}
\usage{
\method{vibe}{gam}(object, varimp = "hp", gof = "R2e", ncores = 1, progress = TRUE, ...)
}
\arguments{
\item{object}{A \code{gam} object, typically result of \code{\link[mgcv:gam]{mgcv::gam()}}.}

\item{varimp}{One of \code{c("hp", "relweights")}, which stand for hierarchical
partitioning and relative weights.}

\item{gof}{Goodness-of-fit metric, the changes of which shall be
analysed}

\item{ncores}{Number of cores used for the model fitting process, happening
in \code{\link{part_core}}.}

\item{progress}{Boolean. Do you want to see a progress bar?}

\item{...}{Other arguments}
}
\description{
\code{vibe.gam} takes a fitted \code{\link[mgcv:gam]{mgcv::gam()}} object and calculates
variable importance metrics by fitting the submodels required, extracting the
desired goodness-of-fit metric and applying variable importance metrics to
it.
}
\examples{
library("mgcv")
gam_ocat <- gam(
  satisfaction ~ admin + hygiene + time_appointment +
    quality_dr + diagnosis_exactness + equipment_modern +
    friendly_workers + parking_playingrooms_cafes,
  data = vibe::sat, family = ocat(R = 3)
)
hp_gam <- vibe(gam_ocat, varimp = "hp", gof = "R2e", progress = FALSE)
rw_gam <- vibe(gam_ocat, varimp = "relweights", gof = "R2e")
print(hp_gam)
print(rw_gam)
}
